requirements.txt
flask>=3.0
google-genai>=1.0
python-dotenv>=1.0
gunicorn>=21.0
tenacity>=8.0
